+++ CHANGELOG.md
@@ -13,10 +13,13 @@ Note: Nextcloud 27 is no longer supported since this version.
 
 - New OCS API endpoint to setAppInitProgress. The old one is marked as deprecated. #319
 - Added default timeout for requestToExApp function set to 3s. #277
+- Added new PublicFunction method `getExApp`. #326
 
 ### Changed
 
 - Dropped support of Nextcloud 27. #322
+- ExApp system flag is now deprecated and removed to optimize performance and simplicity. #323
+- PublicFunctions changes: `exAppRequestWithUserInit` and `asyncExAppRequestWithUserInit` are now deprecated. #323
 
 ### Fixed